shop/requirements.txt
Flask==2.1.1
Flask-RESTful==0.3.9
pymongo==4.1.1
bcrypt==3.2.0
PyJWT==1.7.1
pandas==1.4.2
python-dotenv==0.19.0